prune test hierarchy
