+++ airbyte-integrations/connectors/destination-redshift/src/test-integration/java/io/airbyte/integrations/destination/redshift/SshPasswordRedshiftStagingDestinationAcceptanceTest.java
@@ -4,15 +4,12 @@
 
 package io.airbyte.integrations.destination.redshift;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import io.airbyte.cdk.integrations.base.ssh.SshTunnel.TunnelMethod;
-import io.airbyte.commons.io.IOs;
-import io.airbyte.commons.json.Jsons;
-import java.io.IOException;
-import java.nio.file.Path;
+import static io.airbyte.cdk.integrations.base.ssh.SshTunnel.TunnelMethod.SSH_PASSWORD_AUTH;
+
+import io.airbyte.cdk.integrations.base.ssh.SshTunnel;
 import org.junit.jupiter.api.Disabled;
 
-/*
+/**
  * SshPasswordRedshiftStagingDestinationAcceptanceTest runs basic Redshift Destination Tests using
  * the S3 Staging mechanism for upload of data and "password" authentication for the SSH bastion
  * configuration.
@@ -21,15 +18,8 @@ import org.junit.jupiter.api.Disabled;
 public class SshPasswordRedshiftStagingDestinationAcceptanceTest extends SshRedshiftDestinationBaseAcceptanceTest {
 
   @Override
-  public TunnelMethod getTunnelMethod() {
-    return TunnelMethod.SSH_PASSWORD_AUTH;
-  }
-
-  @Override
-  public JsonNode getStaticConfig() throws IOException {
-    final Path configPath = Path.of("secrets/config_staging.json");
-    final String configAsString = IOs.readFile(configPath);
-    return Jsons.deserialize(configAsString);
+  public SshTunnel.TunnelMethod getTunnelMethod() {
+    return SSH_PASSWORD_AUTH;
   }
 
 }
